refactor: add class attribute to ProfileImage component
