+++ projects/sasil-web/src/components/atoms/ProfileImage/ProfileImage.tsx
@@ -1,26 +1,15 @@
-import styled from '@emotion/styled';
-import { COLORS } from '@sasil/common';
-
-export interface ProfileImageProps {
-  /** 프로필 이미지 url. 사용자가 프로필 이미지를 가지고 있지 않다면 null */
-  imageSrc?: string | null;
-  /** 이미지가 들어가는 원의 width */
-  size: number;
-}
+import * as styles from './ProfileImage.style';
+import { ProfileImageProps } from './ProfileImage.style';
 
 /**
  * 사용자의 프로필 이미지를 나타내는 컴포넌트
  */
-const ProfileImage = styled.div((props: ProfileImageProps) => ({
-  width: props.size,
-  height: props.size,
-  borderRadius: props.size / 2,
-  background: props.imageSrc
-    ? `url(${props.imageSrc})`
-    : COLORS.grayscale.gray6,
-  backgroundRepeat: 'no-repeat',
-  backgroundPosition: 'center center',
-  backgroundSize: 'cover',
-}));
+const ProfileImage = ({ imageSrc, size, className }: ProfileImageProps) => (
+  <styles.styledProfileImage
+    imageSrc={imageSrc}
+    size={size}
+    className={className}
+  />
+);
 
 export default ProfileImage;
