+++ projects/sasil-web/src/components/atoms/ProfileImage/index.ts
@@ -1,5 +1,3 @@
 import ProfileImage from './ProfileImage';
 
 export default ProfileImage;
-
-export type { ProfileImageProps } from './ProfileImage';
